style(debug code): Arrange the debug codes.

+++ include/map.h
@@ -81,7 +81,7 @@ void map_clear_blocks(void);
 int32_t Map_DistanceLimit(uint16_t heading);
 
 int32_t cell_to_count(int16_t distance);
-int16_t count_to_cell(double count);
+int16_t count_to_cell(int32_t count);
 Point32_t map_cell_to_point(const Cell_t& cell);
 Cell_t map_point_to_cell(Point32_t pnt);
